requirements.txt
streamlit
numpy
opencv-python
joblib
Pillow
tensorflow
scikit-learn
